Add norec build tag
